requirements.txt
pytorch
#transformers for the demo